+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### 3.1.1 - 2020-05-15
+
+This fix release contains minor updates and performance improvements.
+
 ## 3.1.0 - May 6, 2020
 
 #### Added